feat(eval_cobalt): wire standalone evaluators as first-class eval_types
Promote the three deterministic evaluators (relevance, grounding, format)
from test-only helpers into first-class eval_types on the eval_cobalt
node:

- Create nodes/src/nodes/eval_cobalt/evaluators/ as a canonical package
  hosting the evaluator functions (pure python, no external deps).
- Add evaluate_relevance, evaluate_grounding, evaluate_format methods
  on CobaltEvaluator and extend the dispatch in evaluate().
- Extend _VALID_EVAL_TYPES and services.json profiles/fields so the
  new eval_types are selectable via the pipeline UI.
- Convert test/cobalt/evaluators/*.py into thin shims that re-export
  from the node package, keeping existing experiment tests unchanged.
- Make eval_cobalt.__init__.py tolerant of a missing rocketlib runtime
  so the evaluators subpackage stays importable in lightweight test
  environments.
- Add unit tests for each of the three new eval_types.
