+++ test/cobalt/evaluators/format_check.py
@@ -21,196 +21,8 @@
 # SOFTWARE.
 # =============================================================================
 
-"""Custom format evaluator for Cobalt experiments.
+"""Shim — re-exports the canonical format evaluator from eval_cobalt node."""
 
-Evaluates whether LLM output matches an expected structural format
-(prose, list, code, json). This evaluator is fully deterministic
-and does not require any external API calls.
-"""
+from eval_cobalt.evaluators.format_check import evaluate_format  # noqa: F401 — re-export for experiments
 
-import json
-import re
-
-
-def _check_prose_format(output: str) -> dict:
-    """Check if output follows prose format (continuous sentences).
-
-    Prose is defined as having at least one sentence-ending punctuation
-    mark and consisting primarily of natural language sentences rather
-    than structured elements like bullet points or code blocks.
-
-    Args:
-        output: The text to check.
-
-    Returns:
-        A dict with score and reasoning.
-    """
-    lines = output.strip().split('\n')
-    non_empty_lines = [line for line in lines if line.strip()]
-
-    if not non_empty_lines:
-        return {'score': 0.0, 'reasoning': 'Output is empty.'}
-
-    # Check for sentence-ending punctuation
-    has_punctuation = bool(re.search(r'[.!?]', output))
-
-    # Check that most lines are not list items
-    list_pattern = re.compile(r'^\s*[-*\u2022]\s|^\s*\d+[.)]\s')
-    list_lines = sum(1 for line in non_empty_lines if list_pattern.match(line))
-    list_ratio = list_lines / len(non_empty_lines) if non_empty_lines else 0
-
-    # Check for code block markers
-    has_code_blocks = '```' in output
-
-    score = 1.0
-    reasons = []
-
-    if not has_punctuation:
-        score -= 0.3
-        reasons.append('Missing sentence-ending punctuation')
-
-    if list_ratio > 0.5:
-        score -= 0.3
-        reasons.append(f'{list_lines}/{len(non_empty_lines)} lines appear to be list items')
-
-    if has_code_blocks:
-        score -= 0.2
-        reasons.append('Contains code blocks')
-
-    score = max(0.0, score)
-    reasoning = '; '.join(reasons) if reasons else 'Output follows prose format'
-
-    return {'score': round(score, 4), 'reasoning': reasoning}
-
-
-def _check_list_format(output: str) -> dict:
-    """Check if output follows list format (bullet points or numbered items).
-
-    Args:
-        output: The text to check.
-
-    Returns:
-        A dict with score and reasoning.
-    """
-    lines = output.strip().split('\n')
-    non_empty_lines = [line for line in lines if line.strip()]
-
-    if not non_empty_lines:
-        return {'score': 0.0, 'reasoning': 'Output is empty.'}
-
-    list_pattern = re.compile(r'^\s*[-*\u2022]\s|^\s*\d+[.)]\s')
-    list_lines = sum(1 for line in non_empty_lines if list_pattern.match(line))
-    list_ratio = list_lines / len(non_empty_lines) if non_empty_lines else 0
-
-    if list_ratio >= 0.5:
-        score = min(1.0, list_ratio + 0.2)
-        reasoning = f'{list_lines}/{len(non_empty_lines)} lines are list items'
-    else:
-        score = list_ratio
-        reasoning = f'Only {list_lines}/{len(non_empty_lines)} lines are list items; expected list format'
-
-    return {'score': round(score, 4), 'reasoning': reasoning}
-
-
-def _check_code_format(output: str) -> dict:
-    """Check if output contains code formatting.
-
-    Looks for code block markers (triple backticks) or consistent
-    indentation patterns typical of code.
-
-    Args:
-        output: The text to check.
-
-    Returns:
-        A dict with score and reasoning.
-    """
-    has_code_blocks = '```' in output
-    lines = output.strip().split('\n')
-    non_empty_lines = [line for line in lines if line.strip()]
-
-    if not non_empty_lines:
-        return {'score': 0.0, 'reasoning': 'Output is empty.'}
-
-    if has_code_blocks:
-        return {'score': 1.0, 'reasoning': 'Output contains code blocks.'}
-
-    # Check for consistent indentation (common in code)
-    indented_lines = sum(1 for line in non_empty_lines if line.startswith('    ') or line.startswith('\t'))
-    indent_ratio = indented_lines / len(non_empty_lines)
-
-    # Check for code-like syntax
-    code_patterns = [r'[{}\[\]();]', r'def\s+\w+', r'class\s+\w+', r'import\s+', r'return\s+', r'=\s*\w+']
-    code_matches = sum(1 for p in code_patterns if re.search(p, output))
-    code_signal = code_matches / len(code_patterns)
-
-    score = max(indent_ratio, code_signal)
-    reasoning = f'Indentation ratio: {indent_ratio:.2f}; Code syntax signals: {code_signal:.2f}'
-
-    return {'score': round(score, 4), 'reasoning': reasoning}
-
-
-def _check_json_format(output: str) -> dict:
-    """Check if output is valid JSON.
-
-    Args:
-        output: The text to check.
-
-    Returns:
-        A dict with score and reasoning.
-    """
-    stripped = output.strip()
-
-    # Try to extract JSON from code blocks
-    code_block_match = re.search(r'```(?:json)?\s*\n(.*?)\n```', stripped, re.DOTALL)
-    if code_block_match:
-        stripped = code_block_match.group(1).strip()
-
-    try:
-        json.loads(stripped)
-        return {'score': 1.0, 'reasoning': 'Output is valid JSON.'}
-    except (json.JSONDecodeError, ValueError) as e:
-        return {'score': 0.0, 'reasoning': f'Output is not valid JSON: {e}'}
-
-
-_FORMAT_CHECKERS = {
-    'prose': _check_prose_format,
-    'list': _check_list_format,
-    'code': _check_code_format,
-    'json': _check_json_format,
-}
-
-
-def evaluate_format(output: str, expected_format: str = 'prose', threshold: float = 0.5) -> dict:
-    """Evaluate output formatting against an expected structural format.
-
-    Supports four format types: prose, list, code, and json.
-
-    Args:
-        output: The LLM-generated output to evaluate.
-        expected_format: The expected format type. One of: prose, list, code, json.
-        threshold: Minimum score to pass (default 0.5).
-
-    Returns:
-        A dict with keys: score (float 0-1), passed (bool), reasoning (str).
-
-    Raises:
-        ValueError: If expected_format is not one of the supported types.
-    """
-    if expected_format not in _FORMAT_CHECKERS:
-        raise ValueError(f'Unsupported format type: {expected_format!r}. Supported: {list(_FORMAT_CHECKERS.keys())}')
-
-    if not output:
-        return {
-            'score': 0.0,
-            'passed': False,
-            'reasoning': f'Output is empty; expected {expected_format} format.',
-        }
-
-    result = _FORMAT_CHECKERS[expected_format](output)
-    passed = result['score'] >= threshold
-
-    return {
-        'score': result['score'],
-        'passed': passed,
-        'reasoning': f'Format: {expected_format}; {result["reasoning"]}; Threshold: {threshold}; Result: {"PASS" if passed else "FAIL"}',
-    }
+__all__ = ['evaluate_format']
